refactor(frameworks): move BaseURLManager implementations into their own packages
Split the five concrete tunnel.BaseURLManager implementations
(Magento2Manager, Magento1Manager, LaravelManager, SymfonyManager,
WordPressManager) out of internal/engine/tunnel/base_url.go into their
owning internal/frameworks/<name>/ packages, so framework-specific tunnel
logic lives alongside each framework's other code instead of a shared
engine package. internal/engine/tunnel/base_url.go now holds only the
BaseURLManager interface and the generic NoopManager fallback.

Batch 1 of the framework-consolidation effort described in
docs/superpowers/specs/2026-08-01-framework-consolidation-design.md.
No behavior change, aside from exporting LaravelManager.UpdateEnv (was
unexported updateEnv) so symfony.SymfonyManager can reach it through
cross-package embedding.
