+++ internal/frameworks/mageos/mageos.go
@@ -24,7 +24,7 @@ func Definition() types.FrameworkDefinition {
 			return NewBootstrap(opts)
 		},
 		BaseURLManager: func() tunnel.BaseURLManager {
-			return &tunnel.Magento2Manager{}
+			return &magento2.Magento2Manager{}
 		},
 		FreshInstall:            freshInstall,
 		PreConfigureHook:        magento2.PreConfigure,
